refactor : onClickListener 설정, 바둑돌 이미지 매퍼 기능 OmokBoardView로 이동
